Updated comments on classes Fixed some issues with the POM
Start to get ready for injecting the JDP into the Service
implementations, in order to do that we need the Set of non-Service
interfaces that the JDP represents.  This is a recursive set of
interfaces.  So if non-Service interface A implements B and C, the A, B
and C would be in this set, and potentially injectable into a Service
implementation.